move inner class
